requirements.txt
streamlit
google-generativeai
PyPDF2
spacy
python-dotenv
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz